[LLZK-221] Move Field, UnreducedInterval, and Interval out of IntervalAnalysis.{cpp,h}
